added alternative HSL picker. Closes #1029

+++ src/game/client/ui.h
@@ -100,7 +100,7 @@ public:
 	float Scale();
 
 	int DoButtonLogic(const void *pID, const char *pText /* TODO: Refactor: Remove */, int Checked, const CUIRect *pRect);
-	int DoPickerLogic(const void *pID, const CUIRect *pRect, int *pX, int *pY);
+	int DoPickerLogic(const void *pID, const CUIRect *pRect, float *pX, float *pY);
 	int DoColorSelectionLogic(const CUIRect *pRect, const CUIRect *pButton);
 
 	// TODO: Refactor: Remove this?
